requirements.txt
wheel
gym-retro
opencv-python
tensorflow-gpu
scipy
numpy
